Moved e2e test to e2e folder
